Create requirements.txt
Here is an extended description of the `requirements.txt` file:

Overview
The `requirements.txt` file is a text file that lists the dependencies required by a Python project. In this case, the file specifies the dependencies required by a Django project.

Dependencies
The file lists the following dependencies:

1. *Django*: The Django framework is specified with a version range of `>=3.0,<4.0`. This means that the project requires Django version 3.0 or higher, but less than version 4.0.
2. *Django Rest Framework*: The Django Rest Framework is specified with a version range of `>=3.11,<4.0`. This means that the project requires Django Rest Framework version 3.11 or higher, but less than version 4.0.
3. *Pillow*: The Pillow library is specified with a version range of `>=7.0,<8.0`. This means that the project requires Pillow version 7.0 or higher, but less than version 8.0.

Purpose
The purpose of the `requirements.txt` file is to specify the dependencies required by the project. This file is used by package managers like pip to install the required dependencies.

Usage
To use this file, simply place it in the root directory of your project. Then, run the following command to install the dependencies:

```
bash
pip install -r requirements.txt
```